Do not use Spring Boot component scan for creating beans
The file based registry, the Mongo DB registry and the Device Connection
service now create service bean instances explicitly in a central place
which makes it easier to understand what's actually
happening.
