+++ service-base/src/main/java/org/eclipse/hono/service/amqp/AbstractDelegatingRequestResponseEndpoint.java
@@ -19,7 +19,6 @@ import java.util.Objects;
 import org.eclipse.hono.config.ServiceConfigProperties;
 import org.eclipse.hono.service.Lifecycle;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 
 import io.vertx.core.Promise;
 import io.vertx.core.Vertx;
@@ -62,7 +61,6 @@ public abstract class AbstractDelegatingRequestResponseEndpoint<S, T extends Ser
      * @throws NullPointerException if service is {@code null}.
      */
     @Autowired
-    @Qualifier("backend")
     public void setService(final S service) {
         Objects.requireNonNull(service);
         log.debug("using service instance: {}", service);
